Branch checkin for variables > 2Gb: tag is "BigVars" because CVS would not allow the tag to start with a digit (e.g. "64bit").
Note that email to 'ncl-dev' was turned off for this checkin.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10079 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/nyacc/reader.c
@@ -1720,7 +1720,8 @@ pack_grammar()
 print_grammar()
 {
     register int i, j, k;
-    int spacing;
+/*    int spacing;*/
+    size_t spacing;
     register FILE *f = verbose_file;
 
     if (!vflag) return;
